+++ BackEnd/BusinessLogic/ChaukaApp.BusinessLogic.csproj
@@ -44,9 +44,11 @@
     <Compile Include="DbConnection.cs" />
     <Compile Include="IEventHost.cs" />
     <Compile Include="IMapPositioner.cs" />
+    <Compile Include="IUserAuthenticator.cs" />
     <Compile Include="MapPositioner.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="EventHost.cs" />
+    <Compile Include="UserAuthenticator.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\DataAccess\ChaukaApp.DataAccess.csproj">
