+++ BackEnd/BusinessLogic/EventHost.cs
@@ -12,40 +12,6 @@
 
         public IEventsRepository<Event> EventRepository { get; set; } = new EventsRepository();
 
-        public IResult<Account> Authentication(string accountName)
-        {
-            IResult<Account> result = new ResultEntity<Account>();
-            result.Success = false;
-            if (this.Repository != null)
-            {
-                if (!string.IsNullOrEmpty(accountName))
-                {
-                    var user = this.Repository.GetUserByAccountName(accountName);
-
-                    if (user != null)
-                    {
-                        result.Success = true;
-                        result.Data = new Account() { Id = user.Id, Name = user.Name };
-                        result.Message = "Successful sign in";
-                    }
-                    else
-                    {
-                        result.Message = "User not found";
-                    }
-                }
-                else
-                {
-                    result.Message = "The account name must not be empty";
-                }
-            }
-            else
-            {
-                result.Message = "It is not possible to access the data service";
-            }
-
-            return result;
-        }
-
         public IResult<UserEvent> GetUserEvents(int userId)
         {
             IResult<UserEvent> eventsResult = new ResultEntity<UserEvent>();
