+++ BackEnd/BusinessLogic/IEventHost.cs
@@ -10,8 +10,6 @@
 
         IEventsRepository<Event> EventRepository { get; set; }
 
-        IResult<Account> Authentication(string accountName);
-
         IResult<UserEvent> GetUserEvents(int userId);
 
         ResultSimplified RegisterEvent(Event newEvent);
